Internal: Rename default styles provider module [ED-22046]
