+++ packages/packages/core/editor-default-styles/src/init.ts
@@ -8,7 +8,7 @@ import { __registerSlice as registerSlice } from '@elementor/store';
 import { PopulateStore } from './components/populate-store';
 import { DefaultStylesOpenGate } from './default-styles-open-gate';
 import { panel } from './default-styles-panel';
-import { defaultStylesStylesProvider } from './default-styles-styles-provider';
+import { defaultStylesStylesProvider } from './default-styles-provider';
 import { useDefaultStylesActionProps } from './hooks/use-default-styles-action-props';
 import { slice } from './store';
